the terrible commit awful commit but i'm not going through this do not use jitter antiaim in its current state

+++ Amalgam/src/SDK/Definitions/Misc/VGUI.h
@@ -1,6 +1,6 @@
 #pragma once
 
-typedef unsigned int VPANEL;
+typedef unsigned long long VPANEL;
 typedef unsigned long HScheme;
 typedef unsigned long HTexture;
 typedef unsigned long HCursor;
